BaseContourGenerator
